addressing clarity comments and refactoring
